tweak test commands
