+++ package.json
@@ -12,11 +12,10 @@
     "link:emails": "rm -rf emails; ln -s src/generator_templates/ts/emails emails;",
     "link:emails:js": "rm -rf emails; ln -s src/generator_templates/js/emails emails;",
     "build": "preconstruct build",
-    "test": "yarn link:emails; jest && yarn test:next-cli",
-    "test:watch": "yarn test --watch",
-    "test:next-cli": "NEXT_CLI_TESTS=1 jest",
-    "ci": "start-server-and-test ci:server http://localhost:3883 test",
-    "ci:server": "cd packages/cli && rm -rf .mailing && src/dev.js --quiet",
+    "test": "yarn link:emails; yarn jest",
+    "test:next": "yarn link:emails; yarn jest -c jest.next.config.ts",
+    "ci:test": "start-server-and-test 'yarn workspaces run mailing ci:server' http://localhost:3883 test",
+    "ci:test:next": "start-server-and-test 'yarn workspaces run mailing ci:server' http://localhost:3883 test:next",
     "prepublish": "yarn build",
     "watch": "preconstruct watch",
     "release": "yarn build && ./scripts/prepare-release && yarn changeset publish",
