+++ packages/cli/package.json
@@ -84,7 +84,8 @@
     "yargs": "^17.5.1"
   },
   "scripts": {
-    "ci:server": "rm -rf .mailing && nohup src/dev.js --quiet > /dev/null 2>&1 &"
+    "ci:server": "rm -rf .mailing && src/dev.js --quiet",
+    "ci:server:nohup": "rm -rf .mailing && nohup src/dev.js --quiet > /dev/null 2>&1 &"
   },
   "keywords": [
     "email",
